acmin-requirements.txt
Django>=2.1.2
requests
python_utils
xlwt
filelock
django-bulk-update
djangorestframework